Started working on the QMap class and its update method to begin Q-learning implementation. Added method visitQ that populates the keys of the Q dictionary as players make their moves. updateQ is now only triggered at the conclusion of a game. For now the plan is, as described in the project proposal, to calculate a reward unit at the end of a winning game as;
r = size**4 / (# of moves made by players in total)

and as r = 0 if the game ended in a tie. Then since every sub-sequence
of the game sequence  is a key in the Q dictionary, its value will
be updated by +r if player1 is winner or -r if player2 is winner. The
intention is that a player using Q for its decisions will try to pick
the next move with maximum value if it is player1 and minimum value if
it is player2, or a random "uncharted" move according to a threshold.