Repeating some changes after I messed up my version control and reverted. The number of iterations I used in the previous description is actually 70 for random and 1000 for reinforcement. I had changed it several times and lost track of the current state. Additionally, I might be able to make the learning work in a way that more closely reflects the process of Q-learning seen in Smartcab. In that case we would update Q by adding a reward plus "the discounted sum of all possible future action states", a process we approximated with stepwise updates of the form;
Q[previous] = reward + gamma * Q[current]

I feel that some of that reasoning can be included in QMap's updateQ method.
Since we update at the very end of the game, in a sense, stepwise, it is as
if we know maybe not all possible future states but all that lead to that
particular game_sequence's end. So I have changed the update be a discounted
sum of the form

Q[subsequence] += sgn * reward * gamma**(k) , for  1< k < length of game_sequence

With this modification the agents still converge to a tied game. I'm starting to
observe games of size 4.