It turns out minimax actually looses to a very small fraction of radom games with its depth set to 2. For example; ((0, 'X'), (4, 'O'), (5, 'X'), (6, 'O'), (2, 'X'), (1, 'O'), (8, 'X')) ((5, 'X'), (4, 'O'), (7, 'X'), (0, 'O'), (8, 'X'), (1, 'O'), (6, 'X')) ((2, 'X'), (4, 'O'), (6, 'X'), (0, 'O'), (8, 'X'), (1, 'O'), (5, 'X'))
When I played it in debug mode using the sequences it had lost to, which
involved forks, and set a depth of 10 and then 4, it would avoid the loss.
I'm starting to think that my implementation is probably not a good one,
which means my measure function is maybe not the best. It still aproaches
perfect play with a depth of 2, never loosing as first player. Using higher
depths in the counts against random playing opponent is prohibitively
slow.