refactor: Begin dividing code base into different files
Although probably not in line with best practices, I am
beginning the process of separating code that is spread in
play.py, game_state.py, and others into a more logical arrangement.
I want to have any code corresponding to tests that I ran in, *tests*
folder. All code that was necessary to produce the figures and reach
conclusions in the report would be in *game_analysis*. Finally the planned
kivy gui and the principal modules would be in the main folder.