feat: Super hacky toggle strategy select screen
In order to get the toggle behavior when selecting a game strategy
I had to pass a reference of SelectList to SelectButton upon
SelectButton creation. This is to mimick the root keyword in kv
language when dealing with nested widgets. It just shows I don't
fully understand how to properly construct with Kivy yet.
The intention is to write this in a more kivy way in the future.
Also game_board.py is now play_game.py.