feat: Have selection screen transition to gameboard grid
Changes are based on the corresponding kivy ScreenManager
tutorial;

https://www.youtube.com/watch?v=xx-NLOg6x8o&list=PLdNh1e1kmiPP4YApJm8ENK2yMlwF1_edq&index=14

Main things to note are that GameBoard and SelectionScreen must
inherit from Screen and a subsequent BoxLayout widget must be added
to them (either through kv language or on the python side). Then all
their properties can be defined as before inside that Boxlayout. Because
of the same reason GameBoard in play_game.py has some hacky lines again
in its init method, where I generate a GridLayout and populate it with
the desired properties as before, and then add it to the parent GameBoard
screen widget.