feat: Sanitize user text input and pass size to GameBoard screen
Though I'm not sure if this is a horrible practice, to add this
functionality I basically cast the text input to an integer and
use the value error thrown when text_input is actually not castable
to prompt the user to input an integer between 2 and 10 exclusive.
The catch statement actually updates the start button to this prompt.
Also I moved all initialization of GameGrid from kivy language to
the python side, as it is directly tied to whether a user gave correct
input in the SelectionScreen. Functions inside SelectScreen take care
of both these steps.