feat: Resolve Q loading logic and the end of game event
I got carried away with this commit and started working on two things
at once (neither is yet finnished). First it is settled now that when
a user selects train-miniQmax the game will be set to learning and
any policy can be pit against train-miniQmax including the perfect
players. The 3x3 pre-trained Q's for minQmax and Q-learning will also
be loaded regardless of the size of the game chosen. In this way the
game will "work" regardless, even if agents will behave stupidly.
Once train-miniQmax has been selected, if there is no previous pickled
Q for a particular size game an empty QMap() is used as a global_Q.
TODO: When a train-miniQmax game ends pickle the resulting Q under
train-miniQmax_<N>X<N>.pickle file name, where <N> is the game size.

Now, I also reworked the clock logic so that I could halt the game
once it finishes. Since I am not too sure how the proper way to handle
this is, what ended up working was to create the Clock instance as
an instance variable of the TicTacScreenManager class coupled with a
global game_has_finished boolean. In this way the clock scheduled event
can be terminated and a popup with a button can be used to return to
the select screen.
TODO: The problem now is to clear the game grid so that the select
screen actually chooses a new game, as it is now it returns to the
previously played game in its final end state.