fix: Have end of game screen switch actually reset game
The main problem that I had when trying to implement this before
is that in kivy there are two ways to handle screen switching.
1) Updating manager.current variable. This literary switches to an
instance already in existence of the other screen.
2) Using manager.switch_to method. Here a new instance of the other
screen is created and the screen which is moved away from is destroyed.
This is actually the behavior I was trying to achieve since once the
game is over the idea is to start a brand new one after selecting
parameters in the selection screen (which can also be destroyed after
use).

Also methods for starting and resetting the game now are housed in
ScreenManager along with the Clock instance. Since its function
is to manage the game screens and control flow, it seemed more
appropriate.