refactor: Organize game choice logic more sensibly
sanitizeGameChoices is now sanitizeTextInput. Its only job is to
make sure the text field has ok input. In this way I can eliminate
the greet function which I had originally intended for a different
purpose. game_size is now a SelectScreen class variable, in this
way it can be set whenever text is input to the text box. In combination
with ready2go which is set to True every time the text input is ok,
the on_press event of the button can now exclusively trigger the game
set up and start of game.