feat: Add human player interaction
The basic way I implemented this was by first having the gameboard
tiles have a reference to their parent widget (in this case Gameboard)
just like the situation for the select screen widgets and buttons.
When it is a human's turn, the game clock is stopped. I added a class
variable to Strateegery, human_move_index. When I tile is clicked,
human_index_movce is updated to the tile's index on the grid, and at
release updateGameboard is called which in turn eventually calls on
Strateegery's human move method (which will return the updated value
of human_move_index along with the player's mark). The clock is also
resumed on release. My gut tells me this is probably the worst design
ever, but I just don't know any better.