feat: Save Q map after train-miniQmax strategy is chosen
By choosing train-miniQmax the miniQmax policy is used but the game
is setup with the 'learning' flag in GameState. Also, in play_game
resetGame is now called from within the screen manager itself, which
is more sensible. As it was before, I encountered a bug where game
would not reset to initial state.